pyproject.toml
[tool.poetry]
name = "stable-diffusion-kanji-generator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.1.0"
transformers = "^4.44.2"
diffusers = "^0.30.3"
pillow = "^10.1.0"
numpy = "^1.26.0"
pycairo = "^1.24.0"
svglib = "^1.5.1"
cairosvg = "^2.7.1"
rich = "^13.8.1"
pytest = "^8.3.3"
torchvision = "^0.19.1"
pyyaml = "^6.0.2"
accelerate = "^0.34.2"
huggingface-hub = "0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"